save good progress; refactored code to have an Objects utility to dynamically create these custom classes i have
